Update methods for downloading test data
Changed from grabbing all ground truth data from github to just the datasets we are using. Before this change, github would only download part of the requested data as github you limit the amount the script would be allowed to download.

Additionally, we now check to see if the specific dataset folder exists and if it is empty before download.