fix: selective download for benchmark datasets
# Fix issues with downloading test data sets from HelmchenLabSoftware repo (#14)

Changed from grabbing all ground truth data from github to just the datasets we are using.
Before this change, the script would only download part of the requested data as github limit the amount the script would be allowed to download.
Additionally, we now check to see if the specific dataset folder exists and if it is empty before download.

---------

Co-authored-by: phildong <35715936+phildong@users.noreply.github.com>